Quality added some functionality

+++ app/src/main/java/com/hub/offershub/fragment/InActiveBusinessFragment.java
@@ -20,6 +20,7 @@ import com.hub.offershub.activity.EditDetailsActivity;
 import com.hub.offershub.adapter.BusinessAdapter;
 import com.hub.offershub.base.BaseFragment;
 import com.hub.offershub.databinding.FragmentInActiveBusinessBinding;
+import com.hub.offershub.dialogfragment.PaymentDialogFragment;
 import com.hub.offershub.listener.CommonListener;
 import com.hub.offershub.model.BusinessModel;
 import com.hub.offershub.model.OfferModel;
@@ -57,15 +58,12 @@ public class InActiveBusinessFragment extends BaseFragment implements View.OnCli
             list.clear();
             page_no = 0;
             binding.swipeRefresh.setRefreshing(false);
-            Log.e("Check_JK", "InActiveShopsFrag List : "+list.size()+" OwnerID : "+AppApplication.getInstance().prefsHelper.getPref(PrefsHelper.ID));
             commonViewModel.getInActiveShops(makeRequest(), myProgressDialog);
         });
         return binding.getRoot();
     }
 
     private void init() {
-        Log.e("Check_JKShop", "init 1 : "+commonViewModel.getMutableInActiveBusiness().hasActiveObservers());
-        Log.e("Check_JKShop", "init 2 : "+commonViewModel.getMutableInActiveBusiness().isInitialized());
         commonViewModel.getInActiveShops(makeRequest(), myProgressDialog);
         getInActiveData();
         getDeleteData();
@@ -100,29 +98,39 @@ public class InActiveBusinessFragment extends BaseFragment implements View.OnCli
     @Override
     public void onItemEdited(Object obj) {
         BusinessModel.Data model = (BusinessModel.Data) obj;
-        OfferModel.Data offerModel = null;
-        Intent i = new Intent(getActivity(), EditDetailsActivity.class);
-        i.putExtra("shop_model", model);
-        i.putExtra("offer_model", offerModel);
-        i.putExtra("isShop", true);
-        startActivity(i);
+        if ("Expired".equals(model.subscription_status)) {
+            if (!paymentDialogFragment.isAdded())
+                paymentDialogFragment.show(getChildFragmentManager(), PaymentDialogFragment.TAG);
+        } else {
+            OfferModel.Data offerModel = null;
+            Intent i = new Intent(getActivity(), EditDetailsActivity.class);
+            i.putExtra("shop_model", model);
+            i.putExtra("offer_model", offerModel);
+            i.putExtra("isShop", true);
+            startActivity(i);
+        }
     }
 
     BusinessModel.Data deleteModel;
     @Override
     public void onItemRemoved(Object obj) {
         deleteModel = (BusinessModel.Data) obj;
-        commonViewModel.deleteShopPermanent(makeDeleteRequest(deleteModel.id), myProgressDialog);
+        if ("Expired".equals(deleteModel.subscription_status)) {
+            if (!paymentDialogFragment.isAdded())
+                paymentDialogFragment.show(getChildFragmentManager(), PaymentDialogFragment.TAG);
+        } else {
+            commonViewModel.deleteShopPermanent(makeDeleteRequest(deleteModel.id), myProgressDialog);
+        }
     }
 
     private void getInActiveData() {
         commonViewModel.getMutableInActiveBusiness().observe(getViewLifecycleOwner(), businessModel -> {
             if (InActiveBusinessFragment.this.getLifecycle().getCurrentState() == Lifecycle.State.RESUMED) {
                 if (businessModel != null) {
-                    Log.e("Check_JK", "InActiveShopsFrag getInActiveData Status : "+businessModel.status);
                     if(businessModel.status.equals("success")) {
                         if (businessModel.data != null) {
                             binding.empty.emptyConstraint.setVisibility(View.GONE);
+                            binding.businessShimmerLayout.setVisibility(View.GONE);
                             binding.businessRecycler.setVisibility(View.VISIBLE);
                             if (page_no == 0) {
                                 if (list.size() > 0)
@@ -197,7 +205,6 @@ public class InActiveBusinessFragment extends BaseFragment implements View.OnCli
     @Override
     public void onPause() {
         super.onPause();
-        Log.e("Check_JKShop", "onPause : "+commonViewModel);
         if (commonViewModel != null) {
             commonViewModel.getMutableInActiveBusiness().removeObservers(getViewLifecycleOwner());
             commonViewModel.getMutableDeleteShopPermanent().removeObservers(getViewLifecycleOwner());
@@ -207,7 +214,6 @@ public class InActiveBusinessFragment extends BaseFragment implements View.OnCli
     @Override
     public void onResume() {
         super.onResume();
-        Log.e("Check_JKShop", "onResume");
         try {
             if (list != null) {
                 if (list.size() > 0)
